refactor: Move TPU CLI validation logic to tpu_info_util and remove tpu_cli_util
