+++ dags/tpu_observability/utils/tpu_info_util.py
@@ -1,11 +1,24 @@
 """Utility for parsing the output of the 'tpu-info' command."""
 
-from dataclasses import dataclass
-from enum import auto
-from enum import IntEnum
+import os
 import re
+import tempfile
+from dataclasses import dataclass
+from enum import auto, Enum, IntEnum
 
 from airflow.decorators import task
+from dags.tpu_observability.utils import jobset_util as jobset
+from dags.tpu_observability.utils import node_pool_util as node_pool
+from dags.tpu_observability.utils import subprocess_util as subprocess
+
+
+class TpuInfoCmd(Enum):
+  """Defines the available tpu-info CLI commands."""
+
+  HELP = "tpu-info -help"
+  VERSION = "tpu-info --version"
+  PROCESS = "tpu-info --process"
+
 
 # A type alias for a parsed row, mapping column headers to their values.
 _TableRow = dict[str, str]
@@ -94,6 +107,34 @@ def parse_tpu_info_output(output: str) -> list[Table]:
     parsed_tables.append(table)
 
   return parsed_tables
+
+
+@task
+def get_tpu_info_from_pod(info: node_pool.Info, pod_name: str) -> str:
+  """
+  Executes the `tpu-info` command in a specified pod and returns its output.
+
+  This task uses kubectl to run the 'tpu-info' command inside the given pod
+  in the 'default' namespace. The output of the command is captured and
+  returned.
+
+  Args:
+    kubeconfig: The path to the kubeconfig file.
+    pod_name: The name of the pod to execute the command in.
+
+  Returns:
+    The standard output from the 'tpu-info' command.
+  """
+  with tempfile.NamedTemporaryFile() as temp_config_file:
+    env = os.environ.copy()
+    env["KUBECONFIG"] = temp_config_file.name
+
+    cmd = " && ".join([
+        jobset.Command.get_credentials_command(info),
+        f"kubectl exec {pod_name} -n default -- tpu-info",
+    ])
+
+    return subprocess.run_exec(cmd, env=env)
 
 
 if __name__ == "__main__":
